Do something for UTF-8 on Windows.
Console support for text encoding on Windows seems a bit all over the map - no doubt there's an Old New Thing post that explains it all.

With a console set to UTF-8 code page, and a MBCS entry point to the console app, there appears to be some lossy conversion and data is lost. With a UTF-16 entry point, things work as hoped after conversion to UTF-8.

Outside the UTF-8 code page, writing to files directly works as expected regardless to console code page. Redirecting via the console is subject to the consoles attempts to convert code pages. If everything is representable in the consoles code page, it seems like that should work (exhaustive testing is infeasible). Otherwise, a lot of replacement characters creep in (?).

Good enough for a simple tool.